Create a ViewQuotePriceDifference component to manage all price difference UI
