+++ ui/app/pages/swaps/view-quote/index.scss
@@ -116,8 +116,6 @@
       }
     }
 
-    
-
     &-contents {
       display: flex;
